Added 4 tests for 2 groups (Sign In and Sign Up)
